test(api): run Outlines admission on Apple CI
